Add support for ZipNN .znn files
Decompress any shard of a model that was previously compressed with
ZipNN. This commit only extends the functionality of load_state_dict(),
making sure to load the model and decompress it as efficiently as
possible by decompressing in chuncks and by avoiding unnecessary I/O
requests.

This is a proof of concept, currently only supporting sharded models
whose index.json been modified to .znn prefexies, safetensors or any
other file. Support for all single files can be readily added. A version
of which can be found in ZipNN's zipnn_hf() patch.